+++ frontend/init.php
@@ -40,7 +40,7 @@ require_once 'epfl-table/controller.php';
 require_once 'epfl-faq/controller.php';
 require_once 'epfl-button/controller.php';
 require_once 'epfl-gallery/controller.php';
-require_once 'epfl-semester-projects/controller.php';
+require_once 'epfl-student-projects/controller.php';
 require_once 'epfl-carousel/controller.php';
 require_once 'epfl-pdf-flipbook/view.php';
 require_once 'epfl-alert/controller.php';
@@ -196,8 +196,8 @@ function wp_gutenberg_epfl_register_blocks() {
 		'render_callback' => 'EPFL\Plugins\Gutenberg\MiniCardDeck\epfl_mini_card_panel_block',
 		));
 
-	register_block_type( 'epfl/semester-projects', array(
-		'render_callback' => 'EPFL\Plugins\Gutenberg\SemesterProjects\epfl_semester_projects_block',
+	register_block_type( 'epfl/student-projects', array(
+		'render_callback' => 'EPFL\Plugins\Gutenberg\StudentProjects\epfl_student_projects_block',
 		));
 
 	register_block_type( 'epfl/carousel', array(
