+++ src/epfl-student-projects/inspector.js
