+++ src/epfl-student-projects/student-projects-icon.js
