+++ src/index.js
@@ -41,7 +41,7 @@ import './epfl-people'
 import './epfl-post-highlight'
 import './epfl-post-teaser'
 import './epfl-quote'
-import './epfl-semester-projects'
+import './epfl-student-projects'
 import './epfl-scheduler'
 import './epfl-social-feed'
 import './epfl-table-filter'
